+++ tests/utils/test_lt.py
@@ -1,6 +1,6 @@
 # test_lt
 
-from frozendict import frozendict
+from frozendict import FrozenDict
 from pytest import raises
 
 from opyprint import dict_lt, lt
@@ -11,9 +11,9 @@ def test_basics():
     assert {'a': 1} != {'a': 2}
     assert {'a': 1} != {'b': 1}
 
-    assert frozendict({'a': 1}) == frozendict({'a': 1})
-    assert frozendict({'a': 1}) != frozendict({'a': 2})
-    assert frozendict({'a': 1}) != frozendict({'b': 1})
+    assert FrozenDict({'a': 1}) == FrozenDict({'a': 1})
+    assert FrozenDict({'a': 1}) != FrozenDict({'a': 2})
+    assert FrozenDict({'a': 1}) != FrozenDict({'b': 1})
 
     assert ('a',) < ('b',)
 
@@ -21,7 +21,7 @@ def test_basics():
         assert {'a': 1} < {'b': 1}
 
     with raises(TypeError):
-        assert frozendict({'a': 1}) < frozendict({'b': 1})
+        assert FrozenDict({'a': 1}) < FrozenDict({'b': 1})
 
     assert isinstance(sorted({'a', 'b'}), list)
     assert sorted({'a', 'c', 'b'}) < sorted({'b', 'c'})
@@ -37,13 +37,13 @@ def test_dict_lt_with_dict():
 
 
 def test_dict_lt_with_frozendict():
-    assert not dict_lt(frozendict({'a': 1}), frozendict({'a': 1}))
-    assert_dict_lt(frozendict({'a': 1}), frozendict({'b': 1}))
-    assert_dict_lt(frozendict({'a': 1}), frozendict({'a': 2}))
+    assert not dict_lt(FrozenDict({'a': 1}), FrozenDict({'a': 1}))
+    assert_dict_lt(FrozenDict({'a': 1}), FrozenDict({'b': 1}))
+    assert_dict_lt(FrozenDict({'a': 1}), FrozenDict({'a': 2}))
 
-    assert_dict_lt(frozendict({'a': 1}), frozendict({'a': 1, 'b': 1}))
-    assert_dict_lt(frozendict({'a': 1, 'b': 1}),
-                   frozendict({'a': 1, 'c': 1}))
+    assert_dict_lt(FrozenDict({'a': 1}), FrozenDict({'a': 1, 'b': 1}))
+    assert_dict_lt(FrozenDict({'a': 1, 'b': 1}),
+                   FrozenDict({'a': 1, 'c': 1}))
 
 
 def assert_dict_lt(obj_1, obj_2):