+++ tests/utils/test_predicates.py
@@ -1,6 +1,6 @@
 # test_predicates
 
-from frozendict import frozendict
+from frozendict import FrozenDict
 import numpy as np
 from opyprint.utils.predicates import (is_bullettable, is_dict, is_multiliner,
                                        is_oneliner, is_set, is_tuple)
@@ -9,8 +9,8 @@ from opyprint.utils.predicates import (is_bullettable, is_dict, is_multiliner,
 def test_is_dict():
     assert is_dict(dict())
     assert is_dict({'a': 1})
-    assert is_dict(frozendict())
-    assert is_dict(frozendict({'a': 1}))
+    assert is_dict(FrozenDict())
+    assert is_dict(FrozenDict({'a': 1}))
 
     assert not is_dict([1, 2, 3])
     assert not is_dict({1, 2, 3})
@@ -29,7 +29,7 @@ def test_is_set():
     assert not is_set(tuple())
     assert not is_set((1, 2, 3))
     assert not is_set({'a': 1})
-    assert not is_set(frozendict({'a': 1}))
+    assert not is_set(FrozenDict({'a': 1}))
     assert not is_set("foobar")
     assert not is_set(123)
 
@@ -43,7 +43,7 @@ def test_is_tuple():
     assert not is_tuple(frozenset({1, 2, 3}))
     assert not is_tuple([1, 2, 3])
     assert not is_tuple({'a': 1})
-    assert not is_tuple(frozendict({'a': 1}))
+    assert not is_tuple(FrozenDict({'a': 1}))
     assert not is_tuple("foobar")
     assert not is_tuple(123)
 
@@ -68,7 +68,7 @@ def test_is_bullettable():
         range(0, 10),
         dict(),
         {'a': 1},
-        frozendict({'a': 1}),
+        FrozenDict({'a': 1}),
     )
     for obj in bullettables:
         assert is_bullettable(obj)